requirements.txt
# ── Core vision & model ──────────────────────────────────────────────────────
ultralytics>=8.2.0          # YOLOv8 — includes torch dependency resolution
# opencv-python-headless is used instead of opencv-python because Colab (and
# most HPC/server environments) have no display server. The headless variant
# is identical except it omits the GUI dependencies.
opencv-python-headless>=4.9.0

# ── Tracking ─────────────────────────────────────────────────────────────────
# deep_sort_realtime wraps DeepSORT with a built-in MobileNet Re-ID extractor.
# It is the most maintained, pip-installable DeepSORT implementation.
deep-sort-realtime>=1.3.2

# ── Configuration & utilities ────────────────────────────────────────────────
PyYAML>=6.0
numpy>=1.24.0
tqdm>=4.66.0

# ── Video download ───────────────────────────────────────────────────────────
yt-dlp>=2024.3.10

# ── Optional: GPU acceleration (PyTorch is pulled by ultralytics) ────────────
# If you need a specific CUDA version install manually:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
